refactor(ES-014): SessionManager から EngineRunner と CronCommandHandler を抽出
- sessions/engine-runner.ts: runSession と mergeTransportMeta を standalone 関数として抽出
- sessions/cron-command-handler.ts: handleCronCommand を standalone 関数として抽出
- sessions/manager.ts: 982行 → 277行（300行以下を達成）

pnpm build && pnpm test: 673 tests PASS
npx madge --circular: 0 件
